cosmosTestnet -> LocalState
