+++ src/testcases/run_in_band/interchain_kv_query.test.ts
@@ -17,399 +17,36 @@ import {
   waitForICQResultWithRemoteHeight,
 } from '@neutron-org/neutronjsplus/dist/icq';
 import { CodeId, NeutronContract } from '@neutron-org/neutronjsplus/dist/types';
-import { paramChangeProposal } from '@neutron-org/neutronjsplus/dist/proposal';
-import axios from 'axios';
 import { msgDelegate, msgUndelegate } from '../../helpers/gaia';
-import { TestStateLocalCosmosTestNet } from '../../helpers/cosmosTestnet';
+import { LocalState } from '../../helpers/localState';
 import {
   createWalletWrapper,
   WalletWrapper,
 } from '@neutron-org/neutronjsplus/dist/wallet_wrapper';
 import { Coin } from '@cosmjs/proto-signing';
 import { msgSubmitProposal, msgVote } from '../../helpers/gaia';
+import {
+  acceptInterchainqueriesParamsChangeProposal,
+  getCosmosSigningInfosResult,
+  getDelegatorUnbondingDelegationsResult,
+  getKvCallbackStatus,
+  getProposalsResult,
+  getProposalVotesResult,
+  getQueryDelegatorDelegationsResult,
+  getValidatorsSigningInfosResult,
+  registerBalanceQuery,
+  registerDelegatorDelegationsQuery,
+  registerGovProposalsQuery,
+  registerProposalVotesQuery,
+  registerSigningInfoQuery,
+  registerUnbondingDelegationsQuery,
+  removeQuery,
+  removeQueryViaTx,
+  validateBalanceQuery,
+  watchForKvCallbackUpdates,
+} from '../../helpers/kvQuery';
 
 const config = require('../../config.json');
-
-const getKvCallbackStatus = (
-  cm: CosmosWrapper,
-  contractAddress: string,
-  queryId: number,
-) =>
-  cm.queryContract<{
-    last_update_height: number;
-  }>(contractAddress, {
-    kv_callback_stats: {
-      query_id: queryId,
-    },
-  });
-
-const watchForKvCallbackUpdates = async (
-  neutronCm: CosmosWrapper,
-  targetCm: CosmosWrapper,
-  contractAddress: string,
-  queryIds: number[],
-) => {
-  const statusPrev = await Promise.all(
-    queryIds.map((i) => getKvCallbackStatus(neutronCm, contractAddress, i)),
-  );
-  const targetHeight = await targetCm.getHeight();
-  await Promise.all(
-    queryIds.map((i) =>
-      waitForICQResultWithRemoteHeight(
-        neutronCm,
-        contractAddress,
-        i,
-        targetHeight,
-      ),
-    ),
-  );
-  const status = await Promise.all(
-    queryIds.map((i) => getKvCallbackStatus(neutronCm, contractAddress, i)),
-  );
-  for (const i in status) {
-    expect(statusPrev[i].last_update_height).toBeLessThan(
-      status[i].last_update_height,
-    );
-  }
-};
-
-const getQueryBalanceResult = (
-  cm: CosmosWrapper,
-  contractAddress: string,
-  queryId: number,
-) =>
-  cm.queryContract<{
-    balances: {
-      coins: {
-        denom: string;
-        amount: string;
-      }[];
-    };
-    last_submitted_local_height: number;
-  }>(contractAddress, {
-    balance: {
-      query_id: queryId,
-    },
-  });
-
-const getValidatorsSigningInfosResult = (
-  cm: CosmosWrapper,
-  contractAddress: string,
-  queryId: number,
-) =>
-  cm.queryContract<{
-    signing_infos: {
-      signing_infos: {
-        address: string;
-        start_height: string;
-        index_offset: string;
-        jailed_until: string;
-        tombstoned: boolean;
-        missed_blocks_counter: number;
-      }[];
-    };
-    last_submitted_local_height: number;
-  }>(contractAddress, {
-    validators_signing_infos: {
-      query_id: queryId,
-    },
-  });
-
-const getDelegatorUnbondingDelegationsResult = (
-  cm: CosmosWrapper,
-  contractAddress: string,
-  queryId: number,
-) =>
-  cm.queryContract<{
-    unbonding_delegations: {
-      unbonding_responses: {
-        delegator_address: string;
-        validator_address: string;
-        entries: {
-          balance: string;
-          completion_time: string | null;
-          creation_height: number;
-          initial_balance: string;
-        }[];
-      }[];
-    };
-    last_submitted_local_height: number;
-  }>(contractAddress, {
-    get_unbonding_delegations: {
-      query_id: queryId,
-    },
-  });
-
-const getCosmosSigningInfosResult = async (sdkUrl: string) => {
-  try {
-    return (await axios.get(`${sdkUrl}/cosmos/slashing/v1beta1/signing_infos`))
-      .data;
-  } catch (e) {
-    return null;
-  }
-};
-
-const getQueryDelegatorDelegationsResult = (
-  cm: CosmosWrapper,
-  contractAddress: string,
-  queryId: number,
-) =>
-  cm.queryContract<{
-    delegations: {
-      delegator: string;
-      validator: string;
-      amount: {
-        denom: string;
-        amount: string;
-      };
-    }[];
-    last_submitted_local_height: number;
-  }>(contractAddress, {
-    get_delegations: {
-      query_id: queryId,
-    },
-  });
-
-const registerBalanceQuery = async (
-  cm: WalletWrapper,
-  contractAddress: string,
-  connectionId: string,
-  updatePeriod: number,
-  denom: string,
-  addr: string,
-) => {
-  const txResult = await cm.executeContract(contractAddress, {
-    register_balance_query: {
-      connection_id: connectionId,
-      denom: denom,
-      addr: addr,
-      update_period: updatePeriod,
-    },
-  });
-
-  const attribute = getEventAttribute(txResult.events, 'neutron', 'query_id');
-
-  const queryId = parseInt(attribute);
-  expect(queryId).toBeGreaterThanOrEqual(0);
-
-  return queryId;
-};
-
-const registerSigningInfoQuery = async (
-  cm: WalletWrapper,
-  contractAddress: string,
-  connectionId: string,
-  updatePeriod: number,
-  valcons: string,
-) => {
-  const txResult = await cm.executeContract(contractAddress, {
-    register_validators_signing_info_query: {
-      connection_id: connectionId,
-      validators: [valcons],
-      update_period: updatePeriod,
-    },
-  });
-
-  const attribute = getEventAttribute(txResult.events, 'neutron', 'query_id');
-
-  const queryId = parseInt(attribute);
-  expect(queryId).toBeGreaterThanOrEqual(0);
-
-  return queryId;
-};
-
-const registerUnbondingDelegationsQuery = async (
-  cm: WalletWrapper,
-  contractAddress: string,
-  connectionId: string,
-  updatePeriod: number,
-  delegator: string,
-  validator: string,
-) => {
-  const txResult = await cm.executeContract(contractAddress, {
-    register_delegator_unbonding_delegations_query: {
-      connection_id: connectionId,
-      delegator,
-      validators: [validator],
-      update_period: updatePeriod,
-    },
-  });
-
-  const attribute = getEventAttribute(txResult.events, 'neutron', 'query_id');
-
-  const queryId = parseInt(attribute);
-  expect(queryId).toBeGreaterThanOrEqual(0);
-
-  return queryId;
-};
-
-const acceptInterchainqueriesParamsChangeProposal = async (
-  cm: WalletWrapper,
-  title: string,
-  description: string,
-  key: string,
-  value: string,
-) => {
-  const daoCoreAddress = (await cm.chain.getChainAdmins())[0];
-  const daoContracts = await getDaoContracts(cm.chain, daoCoreAddress);
-  const dao = new Dao(cm.chain, daoContracts);
-  const daoMember = new DaoMember(cm, dao);
-  const message = paramChangeProposal({
-    title,
-    description,
-    subspace: 'interchainqueries',
-    key,
-    value,
-  });
-  await dao.makeSingleChoiceProposalPass(
-    [daoMember],
-    title,
-    description,
-    [message],
-    '1000',
-  );
-};
-
-const removeQuery = async (
-  cm: WalletWrapper,
-  contractAddress: string,
-  queryId: number,
-) =>
-  await cm.executeContract(
-    contractAddress,
-    {
-      remove_interchain_query: {
-        query_id: queryId,
-      },
-    },
-    [],
-  );
-
-const removeQueryViaTx = async (
-  cm: WalletWrapper,
-  queryId: bigint,
-  sender: string = cm.wallet.address,
-) => await cm.msgRemoveInterchainQuery(queryId, sender);
-
-const registerDelegatorDelegationsQuery = async (
-  cm: WalletWrapper,
-  contractAddress: string,
-  connectionId: string,
-  updatePeriod: number,
-  delegator: string,
-  validators: string[],
-) => {
-  await cm.executeContract(contractAddress, {
-    register_delegator_delegations_query: {
-      delegator: delegator,
-      validators: validators,
-      connection_id: connectionId,
-      update_period: updatePeriod,
-    },
-  });
-};
-
-const validateBalanceQuery = async (
-  neutronCm: CosmosWrapper,
-  targetCm: CosmosWrapper,
-  contractAddress: string,
-  queryId: number,
-  address: string,
-) => {
-  const interchainQueryResult = await getQueryBalanceResult(
-    neutronCm,
-    contractAddress,
-    queryId,
-  );
-  const directQueryResult = await targetCm.queryBalances(address);
-  expect(filterIBCDenoms(interchainQueryResult.balances.coins)).toEqual(
-    filterIBCDenoms(directQueryResult.balances),
-  );
-};
-
-const registerProposalVotesQuery = async (
-  cm: WalletWrapper,
-  contractAddress: string,
-  connectionId: string,
-  updatePeriod: number,
-  proposalId: number,
-  voters: string[],
-) => {
-  const txResult = await cm.executeContract(contractAddress, {
-    register_government_proposal_votes_query: {
-      connection_id: connectionId,
-      update_period: updatePeriod,
-      proposals_ids: [proposalId],
-      voters: voters,
-    },
-  });
-
-  const attribute = getEventAttribute(txResult.events, 'neutron', 'query_id');
-
-  const queryId = parseInt(attribute);
-  expect(queryId).toBeGreaterThanOrEqual(0);
-
-  return queryId;
-};
-
-const getProposalVotesResult = (
-  cm: CosmosWrapper,
-  contractAddress: string,
-  queryId: number,
-) =>
-  cm.queryContract<{
-    votes: {
-      proposal_votes: {
-        proposal_id: number;
-        voter: string;
-        options: any;
-      }[];
-    };
-    last_submitted_local_height: number;
-  }>(contractAddress, {
-    government_proposal_votes: {
-      query_id: queryId,
-    },
-  });
-
-const registerGovProposalsQuery = async (
-  cm: WalletWrapper,
-  contractAddress: string,
-  connectionId: string,
-  updatePeriod: number,
-  proposalsIds: number[],
-) => {
-  const txResult = await cm.executeContract(contractAddress, {
-    register_government_proposals_query: {
-      connection_id: connectionId,
-      update_period: updatePeriod,
-      proposals_ids: proposalsIds,
-    },
-  });
-
-  const attribute = getEventAttribute(txResult.events, 'neutron', 'query_id');
-
-  const queryId = parseInt(attribute);
-  expect(queryId).toBeGreaterThanOrEqual(0);
-
-  return queryId;
-};
-
-const getProposalsResult = (
-  cm: CosmosWrapper,
-  contractAddress: string,
-  queryId: number,
-) =>
-  cm.queryContract<{
-    proposals: {
-      proposals: any[];
-    };
-    last_submitted_local_height: number;
-  }>(contractAddress, {
-    government_proposals: {
-      query_id: queryId,
-    },
-  });
 
 describe('Neutron / Interchain KV Query', () => {
   const connectionId = 'connection-0';
@@ -418,7 +55,7 @@ describe('Neutron / Interchain KV Query', () => {
     3: 4,
     4: 3,
   };
-  let testState: TestStateLocalCosmosTestNet;
+  let testState: LocalState;
   let neutronChain: CosmosWrapper;
   let gaiaChain: CosmosWrapper;
   let neutronAccount: WalletWrapper;
@@ -429,7 +66,7 @@ describe('Neutron / Interchain KV Query', () => {
 
   beforeAll(async () => {
     const mnemonics = inject('mnemonics');
-    testState = new TestStateLocalCosmosTestNet(config, mnemonics);
+    testState = new LocalState(config, mnemonics);
     await testState.init();
     neutronChain = new CosmosWrapper(
       NEUTRON_DENOM,
