+++ src/testcases/run_in_band/parameters.test.ts
@@ -1,4 +1,4 @@
-import { TestStateLocalCosmosTestNet } from './../../helpers/cosmosTestnet';
+import { LocalState } from './../../helpers/localState';
 import '@neutron-org/neutronjsplus';
 import {
   CosmosWrapper,
@@ -27,7 +27,7 @@ import {
 const config = require('../../config.json');
 
 describe('Neutron / Parameters', () => {
-  let testState: TestStateLocalCosmosTestNet;
+  let testState: LocalState;
   let neutronChain: CosmosWrapper;
   let neutronAccount: WalletWrapper;
   let daoMember1: DaoMember;
@@ -35,7 +35,7 @@ describe('Neutron / Parameters', () => {
 
   beforeAll(async () => {
     const mnemonics = inject('mnemonics');
-    testState = new TestStateLocalCosmosTestNet(config, mnemonics);
+    testState = new LocalState(config, mnemonics);
     await testState.init();
     neutronChain = new CosmosWrapper(
       NEUTRON_DENOM,
